refactor(math): extract core base class and de-duplicate math module (#47)
First de-duplication step of the stub/monolith cleanup (#47), guarded by the
57-test parity suite.

- NEW src/core/core.ts: the base jsfeatNext class (constants, per-instance
  cache/data-type helpers, static module slots), extracted verbatim from the
  monolith. Modules can now extend it without a circular import against the
  aggregator — the keystone every subsequent #47 extraction reuses.
- src/math/math.ts: replace the type-only stub (every method threw 'Method
  not implemented') with the REAL implementation moved verbatim from the
  monolith (get_gaussian_kernel, deprecated perspective_4point_transform,
  qsort, median). Only deliberate change: median() now accepts typed arrays
  in its signature — motion_estimator.lmeds always called it with a
  Float32Array cache buffer; the old stub hid that behind 'any'.
- src/jsfeatNext.ts: shrinks by ~560 lines; imports the base from core and
  attaches math from its module (jsfeatNext.math = math).

Temporary (tightened as #47 proceeds): the affine2d/homography2d static
slots on the base are typed 'any' until those classes move out of the
monolith.

Verified behavior-preserving: tsc --noEmit clean; npm test 57/57; UMD
require shape, instanceof chain (new jsfeatNext.math() instanceof
jsfeatNext), static-constant inheritance (jsfeatNext.math.EPSILON) and
gaussian-kernel numerics all checked against the pre-refactor build.
